01.09 haim (1) init

+++ lib/backend/login/login.dart
@@ -152,10 +152,10 @@ level_update() async {
   }
 
   int user_lv = int.parse(user_data['user_lv']);
-  int user_reward = int.parse(user_data['reward']);
+  double user_reward = double.parse(user_data['reward']);
 
-  lv_start = int.parse(leveling[user_lv-1]['reward']);
-  lv_end = int.parse(leveling[user_lv]['reward']);
+  lv_start = double.parse(leveling[user_lv-1]['reward']);
+  lv_end = double.parse(leveling[user_lv]['reward']);
 
   lv_percent = (user_reward - lv_start)/(lv_end-lv_start);
   //print("lv_percent : $lv_percent");
